+++ app/src/main/java/nctu/cs/cgv/itour/service/MyFirebaseMessagingService.java
@@ -36,8 +36,8 @@ public class MyFirebaseMessagingService extends FirebaseMessagingService {
         // send message to activities by broadcasting
         Intent intent = new Intent("checkinIcon");
         intent.putExtra("postId", postId);
-        intent.putExtra("lat", Double.valueOf(lat));
-        intent.putExtra("lng", Double.valueOf(lng));
+        intent.putExtra("lat", Float.valueOf(lat));
+        intent.putExtra("lng", Float.valueOf(lng));
         LocalBroadcastManager.getInstance(this).sendBroadcast(intent);
     }
 }
